+++ mobile/src/pages/coaching-improvment-feedback-edit/coaching-improvment-feedback-edit.ts
@@ -217,4 +217,10 @@ export class CoachingImprovmentFeedbackEditPage implements OnInit {
   private makeNotEmpty(value: string, defaultValue: string): string {
     return value && value.trim().length > 0 ?  value : defaultValue;
   }
+  onSwipe(event) {
+    // console.log('onSwipe', event);
+    if (event.direction === 4) {
+      this.saveNback();
+    }
+  }
 }