+++ release-notes.md
@@ -26,7 +26,6 @@ This file lists the development task to do:
 - Replace link to Coaching edit page by coaching game page
 - Show in row the buttons of period of time in coaching game page
 - Settings local to define the number of period of time during a game
-- Use a slide transition between coaching edit et coaching game
 - Validate the subscription of new user
 - Fix countr from french referee
 - Solution d'import csv de referee
@@ -36,6 +35,8 @@ This file lists the development task to do:
 ## Version 2.1.3
 
 - Show a Loading screen on login
+- Use a slide transition between coaching edit page, coaching game page, positive feedback page and improvment feedback page
+
 
 ## Version 2.1.2  (current)